move boskos to a separate pkg
